Associate label with select element on options page (#552)
* For select elements on the settings page, associate the label with the select element

The select element dropdown for "Position" has a visible label but it's not associated with the select element it's visually connected to. This change wraps the existing visible label in a <label> element and assigns it as the label for the select element.

I'm not too familiar with React, so I used the same {id} configuration found in TextInput.tsx and it appears to work correctly without any unintended consequences.

Alternatively, you could assign aria-label={label} inside of the select tag like this:

```
<select
				aria-label={label}
				value={defaultValue}
				disabled={isDisabled}
				onChange={(event) => {
					onChange(event.target.value as T);
				}}
			>
				{children}
			</select>
```

They both accomplish the same thing but in a slightly different way.

* Fixed formatting prettier was complaining about

* Do the import the right way this time (hopefully)