Update ExcludeKeysSetting.tsx
1. Move focus back to the add button after deleting an exclude keys entry. 

Focus is currently being returned to an invisible spot  instead of the actual button. This change returns it to the add button every time. 

If we return it to the delete button, if a row exists, then a user who clicks twice accidentally might delete an entry they wanted. Moving focus back to the Add button helps protect the user from making a change they can't undo.

useRef was added for this and then called in the handleDeletion() function: `addExcludeKeyButtonRef.current?.focus();`

2. Move the explanation text for exclude keys up so the explanation comes first.

3. Change the icon colors to darker colors that have better contrast for a white or black background. Also change the icon color references to use the variables used elsewhere in settings so they'll remain consistent.

4. Change the icons to the larger version. For the trash can, in particular, this makes the target more visible.

5. Add aria label "Add exclude keys pattern" to the add button so it's clear for screen reader users what the button does.

6. Add aria label to the keys to exclude text field so the purpose of the field is more apparent.

7. Add an aria label to the trashcan that makes it apparent what pattern will be deleted: `aria-label={"Delete pattern " + entry[0]}`