Update toggleHints.ts
Call notifyTogglesStatus() after toggling the hints at a global level. This results in a status toast whether you click the icon or use the keyboard shortcut. For reasons I couldn't track down, the status toast was only being displayed when called with Talon directly.

Displaying the status when the icon is clicked or activated will help users understand what happened if they perform the action accidentally.