Add additional context for screen reader users
Changes the title of the notifications to "Rango notifier:" for screen reader users. When screen readers read out the toast notification, the semantic relationships are lost and all of the text runs together. For the toggle status notification, the result is something like "Rango Now Page Host Global dismiss to close"

This change, combined with the others, help make the readout make more sense for screen reader users without affecting the visual formatting.