Add screen reader enhancements
Utilize the sr-only css class to add additional content to the readout of notifications for screen reader users.

I initially tried to solve for this by adding labels to the icons and unhiding them from screen readers, but the toggle status notification puts the icons on the right and the label on the left whereas the other notifications put the icon on the left and the text on the right. The non-toggle notifications already included the status information in the accessible text. Adding the label to the image made them nonsensical.

Instead, this adds additional helper text after the visual label that's only exposed to screen reader users. getIconType() returns the exact text needed, so I used that in the screen reader output.

The expected result for a toggle status dialog is something like: "Rango notifier: Now hints are unset. Page hints are unset. host hints are disabled. Global hints are enabled. dismiss to close"

For non-toggle notifications, the result is something like "Rango notifier: keyboard clicking enabled"