Improve error message handling in schema
A 'message' can be defined for some (all?) zod objects, including z.number(). By assigning an error message to the 'message' property for each schema item (where relevant), we negate the need for the extra logic in NumberInput.tsx to handle NaN instances. The value of 'message' will be returned instead. When the value is outside of the .refine() range, invalid input is returned. We can also apply the same message to that.

Instead writing the same message twice, errors = {...} was added at the top to store the various error messages.

There might be a better way to do this with a custom error map in zod, but this seems to work quite well.