More refined error output
The previous push added a second iterator which was probably not resource friendly. To get around that, I had to wrap the entire block in another div which then required some additional CSS changes to keep the formatting consistent. 

One benefit of this change is that now the alert boxes will appear immediately below the row where the error is occuring.